+++ src/library/DIPS.Mobile.UI/Components/Chips/Android/OnToggledChangedListener.cs
@@ -1,4 +1,5 @@
 using Android.Widget;
+using DIPS.Mobile.UI.Internal.Logging;
 using Microsoft.Maui.Platform;
 
 namespace DIPS.Mobile.UI.Components.Chips;
@@ -16,6 +17,12 @@ internal class OnToggledChangedListener : Java.Lang.Object, CompoundButton.IOnCh
         m_handler.VirtualView.IsToggled = isChecked;
 
         await Task.Delay(1);
+        
+        if (m_handler.PlatformView is null)
+        {
+            DUILogService.LogError<OnToggledChangedListener>("PlatformView is null, this should not happen, most likely the issue is that the Content is rendered and then the handler is instantly disconnected");
+            return;
+        }
         m_handler.PlatformView.CheckedIconTint = m_handler.VirtualView.TitleColor?.ToDefaultColorStateList();
     }
 }