+++ src/library/DIPS.Mobile.UI/Components/Saving/SaveView/SaveView.cs
@@ -53,9 +53,13 @@ public partial class SaveView : ContentView
     
     private bool DidTapToSave { get; set; }
 
-    protected override void OnHandlerChanged()
+    protected override void OnHandlerChanging(HandlerChangingEventArgs args)
     {
-        base.OnHandlerChanged();
+        base.OnHandlerChanging(args);
+
+        if (args.NewHandler is null)
+            return;
+        
         SetSavingText();
     }