+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [40.0.3]
-- [ChipGroup] Make sure functions does not run when handler is null.
+- Make sure functions does not run when handler is null.
+- Make sure to null check `PlatformView` in some handlers and log if it is.
 
 ## [40.0.2]
 - [ImageCapture] Make sure we handle unchecked orientation data